add typed json fetch and simplify types
