+++ src/components/molecules/content-header/content-header.tsx
@@ -1,16 +1,12 @@
 import styles from './content-header.module.scss';
-import { Content } from '../../../types/content';
-import { Author, Authors } from '../../atoms/authors/authors';
+import { Authors } from '../../atoms/authors/authors';
 import { Descriptors } from '../../atoms/descriptors/descriptors';
 import { Title } from '../../atoms/title/title';
-import { ArticleFlagList, Props as FlagProps } from '../article-flag-list/article-flag-list';
+import { ArticleFlagList } from '../article-flag-list/article-flag-list';
 import { Institution, Institutions } from '../../atoms/institutions/institutions';
+import { MetaData } from '../../../types';
 
-export type ContentHeaderProps = FlagProps & {
-  authors: Author[];
-  doi: string;
-  title: Content;
-};
+export type ContentHeaderProps = Omit<MetaData, 'abstract'>;
 
 const filterInstitutions = (institution: Institution | undefined): institution is Institution => institution !== undefined;
