+++ src/types/index.ts
@@ -1,0 +1,2 @@
+export type { MetaData } from './meta-data';
+export type { PeerReview } from './peer-review';
